Add userinfo support to client (#26)
* First pass at userinfo client

This adds a Userinfo method to the OIDC client. This takes a token, and will
fetch the userinfo if the endpoint is provided.

It fetches this into a standard Claims object, and uses this to validate the
returned subject. This can then be extracted out into whatever format using the
Unmarshal method it provides.

An updated token is returned, this can be used to capture the updated refresh
token if a refresh was needed.

Overall this does feel a little wonky, so I think it'll need some deeper
thought.

* Test and fix refreshing

This makes sure refresh on userinfo functions as intended, and cleans up some
error reporting

* Lint cleanups