Versioned session schema (#21)
* Versioned session schema

To evolve the session data over time, it would be idea if we could associate
them with a version that we can act on at deserialization time. This introduces
a top-level type that just tracks the version, and the session as a raw
message. We can then handle this appropriately.

As part of this, we just move to go JSON serialization rather than protobufs.
This is data only we consume and is used in Go code, so proto added a bunch of
noise for no additional value.

Because we do have existing unversioned and proto-based data out there, code
has been added to detect this and parse the data correctly.

* Update code for new SM interface

* lint

* use google, not gogo