middleware: Remove hard dependency on Gorilla Sessions
Right now using the middleware is hard bound to using gorilla's sessions
library, which makes it annoying to use in any project that isn't already using
that for sessions. This adds a more generic load/save implementation for a
custom type, that consumers can implement with whatever their preferred session
management scheme is.

A implementation of this Interface is provided for gorilla sessions, and it has
been added in a way that is fully backwards compatible - existing code needs no
changes. That has been marked as deprecated though, to indicate that we'll want
to get rid of it at some point.

The existing tests exercise both the interface and the adapter.