Use tink for JWT operations
Tink is a higher level, high-quality crypto library from Google. It fits in to
this package well - it provides well defined JWT operations, but more
importantly it already operates on the concept of a "keyset", which every
consumer of this has had to implement from scratch. Moving to this puts us in a
position where consumers are pushed to use a better keyset management than they
probably are.