Add PKCE Support
This pull request introduces PKCE (Proof Key for Code Exchange) support to the
OIDC client and server library. By adding PKCE, we mitigate the risk of
authorization code interception attacks and adhere to current best practices in
OAuth 2.0 and OIDC security. This is added by default to all CLI token flows
which are inherently public. Other flows can opt-in as appropriate. An option
has been added to the server library to force this for all public clients.