Re-work the client
Previously, we had a client implementation that mostly shadowed the underlying
oauth2 usage, and was split in to three parts - a discovery layer that could
get configs from OIDC discovery, a client that mimic'd the oauth2 client, and a
verifier that could verify JWTs against a set of keys. This was pretty complex
and duplicative, and ended up being leaky anyway. We'd allow passing raw
options to the underlying oauth2 parts, and all three were generally required
to be used together.

This replaces all 3 client interfaces with a "Provider". The provider has the
OIDC bits - discovering information, fetching keysets, and verifying JWTs
against it. For the flow, x/oauth2 should be used. To support this our caching
mechanism and clients have been updated with the new interface.