Rework token/OIDC/JWT handling
Verification was tightly coupled to the OIDC provider. This will get weird when
adding support for OAuth2 discovery and result in some duplication, so create a
new top-level jwt package for the token verification side. This can consume an
abstract provider to discover the relevant info, as well as make standalone and
repeated verification easier.

Switch to go-jose for the verification. Looking at adding DPoP, and Tink's
inability to abstract the signer will get annoying for hardware/restricted key
usage. We also ended up with an API that was more tightly bound to the JWT
implementation than needed, we are now in a place where that is abstracted from
the actual API we expose.

Move standard claims into the JWT package, and be more flexible about custom
claim types. In reality every app is slightly different, so make it easier for
them to define their own claims types.