clitoken: Remove encrypted file cache
This has been around since the beginning, and I'm not sure it is actually used.
It's also mildly unweildy, due to prompting for a password on each op.

There are pretty much always better possibilities now with system keychains or
even using something like the op CLI, so drop it.