Add package for OIDC dynamic client registration
We want to support this, so start a package for it. It is pretty basic right
now, just has the standard types and a client to call it from the discovered
endpoint. The oidccli is updated to be able to exercise it.

Some more thought will be needed around things like the middleware, like how to
we handle things like dynamic clients that might need secret rotation during a
process lifecycle. Or even how we can manage more dynamic secrets in general
(like JWT), in apps that consume this it's just been hacks so far.