Clean up Key interfaces, add validation
The jwt.PublicKeyset interface had functions that existed to support discovery,
but were not used by the `jwt` package. Remove that, and add different key
interfaces for discovery. Discovery now supports both our keyset type, as well
as just providing a source for a JWKS to serve. This lets the signing
implementation decide what is easiest.

Add some validation for keys, to make sure we a) always have the correct public
key types, and b) confirm again that they align with the declared type. go-jose
should also have been doing that for us, but it also allows us to error when
keysets are loaded with bad keys.

Discovery has some tink-specific stuff removed, that can go in the JWKS
implementation if tink is used as the signer. It also has a basic test bought
back.