keychain: Fallback to binary hash for all codesig errors
Currently this doesn't work on Intel CPU Macs. The Go binaries aren't ad-hoc
signed on that platform, and our current error check doesn't handle the actual
error returned. Update it to just fall back to binary hash if any error occurs.
I don't think we're likely to end up in a state here where we flap between the
two, and we test them independently so it's not likely to hide implementation
errors.