Implement a generic claims validator/verifier
This allows for custom claim types to be validated and verified against a
provider instance. This will enable us to support custom claim types, and
generate a more extensive set of standard claims. In addition, this enables the
verification and validation of claims against static keysets if desired, and I
think makes for a cleaner interface than hanging methods off provider.