Generate standard ID token claims, document use
This uses the claims generator to generate a claims set for the standard OIDC
ID token claims. An example of use with a provider is added.

For the example, we use a C&P test provider. We might want to consider a more
formal test provider that library users can use to test against.