dpop: Support verifying and creating x509 bound tokens
This is non-standard, part of an experiment. The idea is to achieve
device-bound proofs with some form of device verification like what is possible
in RFC8705, in environments where mTLS is not possible.

DPoP proofs can be created as a "self-signed" JWT, with the certificate chain
corresponding to the signing key included in the x5c header. Verifiers can be
created with a series of trusted roots, if they are the x5c header presence
will be enforced, and it will be checked that it corresponds to one of these
roots.