dpop: JWT Support
This adds support for encoding and decoding JWT's for DPoP. Where possible we
lean on Tink, but we do need to do custom header parsing to get the proof JWKs,
as it doesn't support custom headers. The encoding side is implemented here to
allow opaque signing keys (e.g SEP), and to set the header appropriately.